Issue 6245 - covscan fixes (#6246)
* Issue 6245 - covscan fixes
Fix issues reported by coverity scan static analyzer

Issue: #6245

Reviewed by: @mreynolds389 (Thanks!)

+++ ldap/servers/slapd/back-ldbm/db-mdb/mdb_layer.c
@@ -811,7 +811,7 @@ dbmdb_copyfile(char *source, char *destination, int overwrite __attribute__((unu
     int dest_fd = -1;
     char *buffer = NULL;
     int return_value = -1;
-    int bytes_to_write = 0;
+    size_t bytes_to_write = 0;
 
     /* malloc the buffer */
     buffer = slapi_ch_malloc(64 * 1024);
@@ -856,12 +856,12 @@ dbmdb_copyfile(char *source, char *destination, int overwrite __attribute__((unu
                 break;
             } else {
                 /* means error */
-                slapi_log_err(SLAPI_LOG_ERR, "dbmdb_copyfile", "Failed to write by \"%s\"; real: %d bytes, exp: %d bytes\n",
+                slapi_log_err(SLAPI_LOG_ERR, "dbmdb_copyfile", "Failed to write by \"%s\"; real: %d bytes, exp: %lu bytes\n",
                               strerror(errno), return_value, bytes_to_write);
                 if (return_value > 0) {
                     bytes_to_write -= return_value;
                     ptr += return_value;
-                    slapi_log_err(SLAPI_LOG_NOTICE, "dbmdb_copyfile", "Retrying to write %d bytes\n", bytes_to_write);
+                    slapi_log_err(SLAPI_LOG_NOTICE, "dbmdb_copyfile", "Retrying to write %lu bytes\n", bytes_to_write);
                 } else {
                     break;
                 }
